fix: check for existing headers when setting response headers
